Wave 3: the install operation moves to the owned-daemon module
The owner's wave-3 panel (gpt-5.6-sol agentic + claude-fable-5 +
gemini-3.1-pro-preview) returned two seats clean; gemini independently raised
the same boundary point terra raised in round 2 — installer machinery living
in a gateway module contradicts the thin-proxy rule. Two independent models
converging on unchanged code outweighs the round-1 scope PASS this had been
held against, so the rejection flips: is_immediate_missing_cli_job,
install_missing_harness_cli, the stdout drain and the receipt validator now
live in claudexor_daemon.py — the owned data plane whose custody they already
borrowed — and the accounts proxy only invokes them and translates the typed
result. Mechanical move: no behavior change, tests re-pointed, the custody
allowlist entry follows the Popen, the accounts module drops back out of the
size-ratchet band (entry removed on regen).

The scope seat's remaining critical — extract a Windows npm toolchain so
Connect can install on win32 — is REJECTED with the engine's own contract as
evidence: published Claudexor 3.8.0 refuses `--target local` on Windows
(localPlatformRefusal, typed unsupported_platform), so the demanded extraction
would dead-end one layer lower. The Windows lane stays daemon-only, disclosed;
behavior there is unchanged from before the feature.
