+++ docs/ARCHITECTURE.md
@@ -224,7 +224,7 @@ server.py (Starlette+uvicorn) ← HTTP + WebSocket on configurable host:port (de
       │   ├── extensions.py    ← extensions/skills HTTP surface (GET /api/extensions, GET /api/extensions/<skill>/manifest, ALL /api/extensions/<skill>/<rest:path>, POST /api/skills/<skill>/toggle, POST /api/skills/<skill>/delete, POST /api/skills/<skill>/review, POST /api/skills/<skill>/grants)
       │   ├── marketplace.py   ← ClawHub + OuroborosHub HTTP surface
       │   ├── mcp.py           ← MCP Settings API surface backed by the shared MCPManager
-      │   ├── claudexor_accounts.py ← (D30) Agent accounts HTTP surface (Settings → Agents → Accounts): six thin proxies — counted by HANDLER, and two of them serve more than one route or action — of the owned daemon's account truth (GET /api/claudexor/status[?include=models] — side-effect-free daemon/runtime state + harness catalog + credential profiles with honest verification statuses + quota windows, each facet stamped with its read state, plus the additive `unified_accounts` feature fact read from the engine's own /v2/operations catalog (`get:account-pools` present = the unified account model; unreadable catalog fails closed to the legacy rendering); POST /api/claudexor/wake — owner-initiated daemon start behind Refresh; POST /api/claudexor/login — one Connect intent that installs/repairs the managed runtime, starts or attaches the owned daemon, then creates or re-adopts its setup job; when that exact pinned engine's immediate create response is the pre-authorization structural `not_supported` missing-binary shape, Connect runs the exact managed CLI once as `harness install <harness> --target local --yes --json`, admits only its strict success receipt, and creates the same login job exactly once more; GET/DELETE /api/claudexor/login/{job_id} — canonical snapshot/cancel; POST /api/claudexor/login/{job_id}/input — the owner's answer to a waiting engine prompt; POST /api/claudexor/login/{job_id}/reconcile — an explicit proof-of-empty check after `termination_unconfirmed`; DELETE/PATCH /api/claudexor/credential-profiles/{harness}/{profile_id} — one route, two row actions: ask the engine to forget one named account, or toggle its `enabled` participation in the rotation pool with the engine's own strict `{enabled}` body passed through). Zero auth logic and zero vendor recipes live here, and the browser never sees the daemon token. Create/cancel/input/reconcile wrap one bare job exactly once; snapshot passes the daemon's `{job,cursor,sequence,deviceCode?}` envelope through verbatim, so transient disclosure remains at envelope level instead of being hidden by the former `job.job` shape. A terminal job is not automatically release proof: unreconciled `termination_unconfirmed` retains daemon custody until reconciliation records `status=empty`, while poll/cancel/reconcile 404/410 release only the missing client record. `harness_login_cards.jobDetail()` reads the canonical envelope's one job and renders its escaped, untruncated message only beside a settled non-success verdict, never beside `Connected.` or while the live status line owns the card.
+      │   ├── claudexor_accounts.py ← (D30) Agent accounts HTTP surface (Settings → Agents → Accounts): six thin proxies — counted by HANDLER, and two of them serve more than one route or action — of the owned daemon's account truth (GET /api/claudexor/status[?include=models] — side-effect-free daemon/runtime state + harness catalog + credential profiles with honest verification statuses + quota windows, each facet stamped with its read state, plus the additive `unified_accounts` feature fact read from the engine's own /v2/operations catalog (`get:account-pools` present = the unified account model; unreadable catalog fails closed to the legacy rendering); POST /api/claudexor/wake — owner-initiated daemon start behind Refresh; POST /api/claudexor/login — one Connect intent that installs/repairs the managed runtime, starts or attaches the owned daemon, then creates or re-adopts its setup job; when that exact pinned engine's immediate create response is the pre-authorization structural `not_supported` missing-binary shape, Connect invokes the owned-daemon module's `install_missing_harness_cli` (claudexor_daemon.py owns the whole operation — spawn, custody, receipt policy; this proxy only translates its typed result) and creates the same login job exactly once more; GET/DELETE /api/claudexor/login/{job_id} — canonical snapshot/cancel; POST /api/claudexor/login/{job_id}/input — the owner's answer to a waiting engine prompt; POST /api/claudexor/login/{job_id}/reconcile — an explicit proof-of-empty check after `termination_unconfirmed`; DELETE/PATCH /api/claudexor/credential-profiles/{harness}/{profile_id} — one route, two row actions: ask the engine to forget one named account, or toggle its `enabled` participation in the rotation pool with the engine's own strict `{enabled}` body passed through). Zero auth logic and zero vendor recipes live here, and the browser never sees the daemon token. Create/cancel/input/reconcile wrap one bare job exactly once; snapshot passes the daemon's `{job,cursor,sequence,deviceCode?}` envelope through verbatim, so transient disclosure remains at envelope level instead of being hidden by the former `job.job` shape. A terminal job is not automatically release proof: unreconciled `termination_unconfirmed` retains daemon custody until reconciliation records `status=empty`, while poll/cancel/reconcile 404/410 release only the missing client record. `harness_login_cards.jobDetail()` reads the canonical envelope's one job and renders its escaped, untruncated message only beside a settled non-success verdict, never beside `Connected.` or while the live status line owns the card.
       │   ├── host_service.py  ← Loopback-only Host Service API for reviewed skill callbacks
       │   ├── history.py       ← Chat history + cost breakdown endpoint factories
       │   ├── projects.py      ← Multi-project CRUD surface (v6.32.0): GET /api/projects, POST /api/projects, POST /api/projects/from-task (bind an existing task to a new project). (v6.33.0 removed the /sleep + /wake status endpoints.)
