+++ tests/test_claudexor_owned_daemon.py
@@ -957,7 +957,7 @@ def test_immediate_missing_cli_trigger_requires_exact_job_and_engine(monkeypatch
     from types import SimpleNamespace
 
     from ouroboros import claudexor_runtime as runtime
-    from ouroboros.gateway.claudexor_accounts import _is_immediate_missing_cli_job
+    from ouroboros.claudexor_daemon import is_immediate_missing_cli_job as _is_immediate_missing_cli_job
 
     pin = SimpleNamespace(
         version="3.4.0", build_sha="1" * 40, cli_entrypoint="claudexor.bundle.cjs")
@@ -1017,7 +1017,7 @@ def test_login_installs_and_retries_exactly_once(monkeypatch, retry_job):
             return _missing_cli_job() if len(requests) == 1 else retry_job
 
     monkeypatch.setattr(owned, "ensure_owned_gateway", lambda: Gateway())
-    monkeypatch.setattr(accounts, "_install_harness_cli", installs.append)
+    monkeypatch.setattr(owned, "install_missing_harness_cli", installs.append)
     answer = accounts._login_create({"harness": "cursor"})
 
     assert installs == ["cursor"]
@@ -1026,7 +1026,7 @@ def test_login_installs_and_retries_exactly_once(monkeypatch, retry_job):
 
 
 def test_install_success_receipt_is_strict_and_provenance_is_a_pair():
-    from ouroboros.gateway.claudexor_accounts import _valid_install_success
+    from ouroboros.claudexor_daemon import _valid_install_success
 
     assert _valid_install_success(_install_receipt(), "cursor") is True
     assert _valid_install_success(_install_receipt(
@@ -1075,7 +1075,6 @@ def test_installer_invocation_is_exact_grouped_and_stdout_bounded(monkeypatch):
     from types import SimpleNamespace
 
     from ouroboros import claudexor_runtime as runtime
-    from ouroboros.gateway import claudexor_accounts as accounts
     from ouroboros import platform_layer
 
     command = ["/exact/node", "/exact/closure/claudexor.bundle.cjs"]
@@ -1103,8 +1102,8 @@ def test_installer_invocation_is_exact_grouped_and_stdout_bounded(monkeypatch):
         seen["lock_held_during_spawn"] = shell._subprocess_lock.locked()
         return Process()
 
-    monkeypatch.setattr(accounts.subprocess, "Popen", popen)
-    accounts._install_harness_cli("cursor")
+    monkeypatch.setattr(owned.subprocess, "Popen", popen)
+    owned.install_missing_harness_cli("cursor")
     assert seen["argv"] == [
         *command, "harness", "install", "cursor",
         "--target", "local", "--yes", "--json",
@@ -1123,9 +1122,9 @@ def test_installer_invocation_is_exact_grouped_and_stdout_bounded(monkeypatch):
     assert "CLAUDEXOR_CONTROL_PORT" not in env
     assert seen["lock_held_during_spawn"] is True
     output, state = bytearray(), {}
-    accounts._drain_installer_stdout(
-        io.BytesIO(b"x" * (accounts._HARNESS_INSTALL_STDOUT_LIMIT + 1)), output, state)
-    assert len(output) == accounts._HARNESS_INSTALL_STDOUT_LIMIT
+    owned._drain_installer_stdout(
+        io.BytesIO(b"x" * (owned._HARNESS_INSTALL_STDOUT_LIMIT + 1)), output, state)
+    assert len(output) == owned._HARNESS_INSTALL_STDOUT_LIMIT
     assert state == {"overflow": True}
 
 
@@ -1135,7 +1134,6 @@ def test_installer_timeout_kills_the_process_tree(monkeypatch):
     from types import SimpleNamespace
 
     from ouroboros import claudexor_runtime as runtime
-    from ouroboros.gateway import claudexor_accounts as accounts
     from ouroboros.gateways.claudexor import ClaudexorUnavailable
     from ouroboros import platform_layer
 
@@ -1156,13 +1154,13 @@ def test_installer_timeout_kills_the_process_tree(monkeypatch):
             return -9
 
     proc = Process()
-    monkeypatch.setattr(accounts.subprocess, "Popen", lambda *_a, **_kw: proc)
+    monkeypatch.setattr(owned.subprocess, "Popen", lambda *_a, **_kw: proc)
     killed = []
     import ouroboros.tools.shell as shell
 
     monkeypatch.setattr(shell, "_kill_process_group", killed.append)
     with pytest.raises(ClaudexorUnavailable) as excinfo:
-        accounts._install_harness_cli("cursor")
+        owned.install_missing_harness_cli("cursor")
     assert excinfo.value.code == "harness_install_timeout"
     assert killed == [proc]
     from ouroboros.config import get_claudexor_harness_install_timeout_sec
@@ -1180,7 +1178,6 @@ def test_installer_child_is_panic_tracked_for_its_whole_run(monkeypatch):
     from types import SimpleNamespace
 
     from ouroboros import claudexor_runtime as runtime
-    from ouroboros.gateway import claudexor_accounts as accounts
     from ouroboros import platform_layer
     import ouroboros.tools.shell as shell
 
@@ -1199,8 +1196,8 @@ def test_installer_child_is_panic_tracked_for_its_whole_run(monkeypatch):
             return 0
 
     proc = Process()
-    monkeypatch.setattr(accounts.subprocess, "Popen", lambda *_a, **_kw: proc)
-    accounts._install_harness_cli("cursor")
+    monkeypatch.setattr(owned.subprocess, "Popen", lambda *_a, **_kw: proc)
+    owned.install_missing_harness_cli("cursor")
     assert observed["tracked_during_wait"] is True, "panic sweep must see the live installer"
     assert proc not in shell._active_subprocesses, "custody ends with the request"
